+++ 00-PROJECT/MiniMart/backend/src/main/java/com/backend/product/dto/res/ProductDTO.java
@@ -1,5 +1,7 @@
 package com.backend.product.dto.res;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,8 +15,8 @@ public class ProductDTO {
     private long id;
     private String name;
     private String imageUrl;
-    private Double salePrice;
-    private double compareAtPrice;
+    private BigDecimal salePrice;
+    private BigDecimal compareAtPrice;
     private int stock;
 
     @Override
